Implement collision detection
- Entities now detect collision with other entities and collidable tiles
- Entity positions updated, now center in square rather than the bottom
  left corner. Renderer adjusted to match this. (Was left over from early
  testing)

+++ core/src/java/com/bau5/coalescence/entities/GameEntity.java
@@ -36,7 +36,7 @@ public abstract class GameEntity extends Entity {
         this.playback = new LinkedList<>();
     }
 
-    public GameEntity(int x, int y, int w, int h, Color color) {
+    public GameEntity(float x, float y, int w, int h, Color color) {
         this(new PositionComponent(x, y), new AttributeComponent(w, h, color));
     }
