+++ core/src/java/com/bau5/coalescence/entities/actions/MoveAction.java
@@ -17,7 +17,7 @@ public class MoveAction extends Action {
     }
 
     public MoveAction(Vector2 vec) {
-        this((int)vec.x, (int)vec.y);
+        this(vec.x, vec.y);
     }
 
     @Override
